+++ code/obj_organic.h
@@ -27,8 +27,8 @@ class TOrganic : public TObj {
     virtual void describeObjectSpecifics(const TBeing *) const;
     virtual bool splitMe(TBeing *, const char *);
     virtual void lightMe(TBeing *, silentTypeT);
-    virtual int  sellPrice(int, int, int *);
-    virtual int  shopPrice(int, int, int, int *) const;
+    virtual int  sellPrice(int, float, int *);
+    virtual int  shopPrice(int, int, float, int *) const;
     virtual void buyMe(TBeing *, TMonster *, int, int);
     virtual void sellMe(TBeing *ch, TMonster *, int);
     virtual void valueMe(TBeing *, TMonster *, int);
